Split out hexagon map/grid visuals and navigation into two separate classes
